Add view and start re-writing server.js
Add views folder (one view and scss/css styles) to serve root page with
instructions. Timestamp microserver used as base, note that the URLs for
Heroku still need to be updated once ready to push.

Add key for mongoURL (using a Sandbox at MLab), add environment check for
value of key (local version stored in keys.js in a module.export object, which
is not stored in this git).